sync on 2024/11/13, rev 74a089e04b4efe3b67b0d61cff4b7b6d10974174

+++ prog/engine/anim/animGraph.cpp
@@ -45,7 +45,7 @@ struct AnimBlender::NodeSamplers
   };
 
   Entry samplers[MAX_ANIMS_IN_NODE];
-  uint8_t totalNum : 8;
+  uint32_t totalNum : 8;
   int32_t lastBnl : 24;
 };
